+++ src/Data/Singletons/ShowSing.hs
@@ -29,7 +29,6 @@ module Data.Singletons.ShowSing (
   ShowSing
   ) where
 
-import Data.Kind
 import Data.Singletons.Internal
 import Data.Singletons.Prelude.Instances
 import Data.Singletons.Single
@@ -70,7 +69,7 @@ import qualified GHC.TypeNats as TN
 -- @
 --
 -- Because that quantified constraint is somewhat lengthy, we provide the
--- 'ShowSing' type synonym as a convenient shorthand. Thus, the above instance
+-- 'ShowSing' class synonym as a convenient shorthand. Thus, the above instance
 -- is equivalent to:
 --
 -- @
@@ -87,7 +86,78 @@ import qualified GHC.TypeNats as TN
 -- * A 'Show' instance for the singleton type
 --
 -- What a bargain!
-type ShowSing k = (forall z. Show (Sing (z :: k)) :: Constraint)
+class    (forall (z :: k). Show (Sing z)) => ShowSing k
+instance (forall (z :: k). Show (Sing z)) => ShowSing k
+
+{-
+Note [ShowSing must be a class, not a type synonym]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+It is very, very tempting to define ShowSing as a type synonym, like this:
+
+  type ShowSing k = (forall (z :: k). Show (Sing z) :: Constraint)
+
+In fact, version 2.5 of the library made the mistake of doing this. However,
+we have since learned the folly of our ways and now define ShowSing as class
+with a single instance.
+
+Why shouldn't you define ShowSing as a type synonym? Recursive data types.
+Consider the following example (from #371):
+
+  $(singletons [d|
+    data X a = X1 | X2 (Y a) deriving Show
+    data Y a = Y1 | Y2 (X a) deriving Show
+    |])
+
+This will generate the following instances:
+
+  deriving instance ShowSing (Y a) => Show (Sing (z :: X a))
+  deriving instance ShowSing (X a) => Show (Sing (z :: Y a))
+
+So far, so good. Now, suppose you try to actually `show` a singleton for X.
+For example:
+
+  show (sing @(X1 :: X Bool))
+
+Somewhat surprisingly, this will be rejected by the typechecker with the
+following error:
+
+    • Reduction stack overflow; size = 201
+      When simplifying the following type: Show (Sing z)
+
+To see why this happens, observe what goes on if we expand the occurrences of
+the ShowSing type synonym in the generated instances:
+
+  deriving instance (forall z. Show (Sing (z :: Y a))) => Show (Sing z :: X a)
+  deriving instance (forall z. Show (Sing (z :: X a))) => Show (Sing z :: Y a)
+
+Due to the way QuantifiedConstraints works, GHC will always favor local,
+quantified constraints in the instance contexts over top-level instances.
+Notice that there is both a top-level instance for (Y a) as well as a local,
+quantified instance for (Y a) in scope in the Show instance for (X a), so GHC
+will favor the local instance during instance resolution. But when resolving
+the local instance for (Y a), we are back in the same situation: there is both
+a top-level and local instance for (X a), so the local instance is picked.
+When resolving /that/ instance... we repeat the process and find ourselves in
+an infinite loop that eventually overflows the reduction stack. Eep.
+
+There are two possible ways to solve this problem:
+
+1. Make derived instances' type inference more clever. If you look closely,
+   you'll notice that the `ShowSing (X a)`/`ShowSing (Y a)` constraints in
+   the generated instances are entirely redundant and could safely be left
+   off. But determining this would require significantly improving singletons'
+   Template Haskell capabilities for type inference, which is a path that we
+   usually spurn in favor of keeping the generated code dumb but predictable.
+2. Define `ShowSing` as a class (with a single instance) instead of a type
+   synonym. `ShowSing`-as-a-class ties the recursive knot during instance
+   resolution and thus avoids the problems that the type synonym version
+   suffers from. As an added bonus, it also avoids the need for clients to
+   enable `QuantifiedConstraints` in every module that singles a derived
+   `Show` instance.
+
+Given the two options, (2) is by far the easier option, so that is what we
+ultimately went with.
+-}
 
 ------------------------------------------------------------
 -- TypeLits instances